addressed comments. Moved common code into an include file
